+++ src/main/java/com/tale/controller/admin/ArticleController.java
@@ -8,13 +8,13 @@ import com.blade.mvc.annotation.*;
 import com.blade.mvc.http.Request;
 import com.blade.mvc.ui.RestResponse;
 import com.tale.controller.BaseController;
-import com.tale.dto.LogActions;
-import com.tale.dto.Types;
+import com.tale.model.dto.LogActions;
+import com.tale.model.dto.Types;
 import com.tale.exception.TipException;
 import com.tale.extension.Commons;
-import com.tale.model.Contents;
-import com.tale.model.Metas;
-import com.tale.model.Users;
+import com.tale.model.entity.Contents;
+import com.tale.model.entity.Metas;
+import com.tale.model.entity.Users;
 import com.tale.service.ContentsService;
 import com.tale.service.LogService;
 import com.tale.service.MetasService;
