+++ src/main/java/com/tale/controller/admin/AuthController.java
@@ -13,10 +13,10 @@ import com.blade.mvc.http.Response;
 import com.blade.mvc.http.Session;
 import com.blade.mvc.ui.RestResponse;
 import com.tale.controller.BaseController;
-import com.tale.dto.LogActions;
+import com.tale.model.dto.LogActions;
 import com.tale.exception.TipException;
 import com.tale.init.TaleConst;
-import com.tale.model.Users;
+import com.tale.model.entity.Users;
 import com.tale.service.LogService;
 import com.tale.service.UsersService;
 import com.tale.utils.TaleUtils;
