+++ src/main/java/com/tale/controller/admin/CategoryController.java
@@ -9,15 +9,13 @@ import com.blade.mvc.http.HttpMethod;
 import com.blade.mvc.http.Request;
 import com.blade.mvc.ui.RestResponse;
 import com.tale.controller.BaseController;
-import com.tale.dto.MetaDto;
-import com.tale.dto.Types;
+import com.tale.model.dto.MetaDto;
+import com.tale.model.dto.Types;
 import com.tale.exception.TipException;
 import com.tale.init.TaleConst;
 import com.tale.service.MetasService;
 import com.tale.service.SiteService;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.List;
