+++ src/main/java/com/tale/controller/admin/LinksController.java
@@ -5,14 +5,12 @@ import com.blade.mvc.annotation.*;
 import com.blade.mvc.http.Request;
 import com.blade.mvc.ui.RestResponse;
 import com.tale.controller.BaseController;
-import com.tale.dto.Types;
+import com.tale.model.dto.Types;
 import com.tale.exception.TipException;
-import com.tale.model.Metas;
+import com.tale.model.entity.Metas;
 import com.tale.service.MetasService;
 import com.tale.service.SiteService;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.List;
