+++ src/main/java/com/tale/controller/admin/ThemeController.java
@@ -8,8 +8,8 @@ import com.blade.mvc.annotation.*;
 import com.blade.mvc.http.Request;
 import com.blade.mvc.ui.RestResponse;
 import com.tale.controller.BaseController;
-import com.tale.dto.LogActions;
-import com.tale.dto.ThemeDto;
+import com.tale.model.dto.LogActions;
+import com.tale.model.dto.ThemeDto;
 import com.tale.exception.TipException;
 import com.tale.extension.Commons;
 import com.tale.init.TaleConst;
