+++ src/main/java/com/tale/extension/Theme.java
@@ -2,12 +2,12 @@ package com.tale.extension;
 
 import com.blade.jdbc.model.Paginator;
 import com.blade.kit.StringKit;
-import com.tale.dto.Comment;
-import com.tale.dto.MetaDto;
-import com.tale.dto.Types;
+import com.tale.model.dto.Comment;
+import com.tale.model.dto.MetaDto;
+import com.tale.model.dto.Types;
 import com.tale.init.TaleConst;
-import com.tale.model.Comments;
-import com.tale.model.Contents;
+import com.tale.model.entity.Comments;
+import com.tale.model.entity.Contents;
 import com.tale.service.SiteService;
 import com.tale.utils.TaleUtils;
 import jetbrick.template.runtime.InterpretContext;
