+++ src/main/java/com/tale/init/WebContext.java
@@ -13,12 +13,12 @@ import com.blade.kit.StringKit;
 import com.blade.mvc.view.template.JetbrickTemplateEngine;
 import com.tale.controller.BaseController;
 import com.tale.controller.admin.AttachController;
-import com.tale.dto.Types;
+import com.tale.model.dto.Types;
 import com.tale.extension.AdminCommons;
 import com.tale.extension.Commons;
 import com.tale.extension.JetTag;
 import com.tale.extension.Theme;
-import com.tale.model.ExtSql2o;
+import com.tale.model.entity.ExtSql2o;
 import com.tale.service.OptionsService;
 import com.tale.service.SiteService;
 import com.tale.utils.RewriteUtils;
