+++ src/main/java/com/tale/model/dto/ThemeDto.java
@@ -1,10 +1,13 @@
-package com.tale.dto;
+package com.tale.model.dto;
+
+import lombok.Data;
 
 import java.io.Serializable;
 
 /**
  * Created by biezhi on 2017/3/15.
  */
+@Data
 public class ThemeDto implements Serializable {
 
     /**
@@ -19,22 +22,6 @@ public class ThemeDto implements Serializable {
 
     public ThemeDto(String name) {
         this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public boolean isHasSetting() {
-        return hasSetting;
-    }
-
-    public void setHasSetting(boolean hasSetting) {
-        this.hasSetting = hasSetting;
     }
 
 }
