+++ src/main/java/com/tale/model/dto/Types.java
@@ -1,4 +1,4 @@
-package com.tale.dto;
+package com.tale.model.dto;
 
 /**
  * Created by biezhi on 2017/2/21.
