+++ src/main/java/com/tale/service/LogService.java
@@ -1,6 +1,6 @@
 package com.tale.service;
 
-import com.tale.model.Logs;
+import com.tale.model.entity.Logs;
 
 import java.util.List;
