+++ src/main/java/com/tale/service/impl/CommentsServiceImpl.java
@@ -8,10 +8,10 @@ import com.blade.jdbc.model.Paginator;
 import com.blade.kit.BladeKit;
 import com.blade.kit.DateKit;
 import com.blade.kit.StringKit;
-import com.tale.dto.Comment;
+import com.tale.model.dto.Comment;
 import com.tale.exception.TipException;
-import com.tale.model.Comments;
-import com.tale.model.Contents;
+import com.tale.model.entity.Comments;
+import com.tale.model.entity.Contents;
 import com.tale.service.CommentsService;
 import com.tale.service.ContentsService;
 import com.tale.utils.TaleUtils;
