+++ src/main/java/com/tale/service/impl/ContentsServiceImpl.java
@@ -8,10 +8,10 @@ import com.blade.jdbc.model.PageRow;
 import com.blade.jdbc.model.Paginator;
 import com.blade.kit.DateKit;
 import com.blade.kit.StringKit;
-import com.tale.dto.Types;
+import com.tale.model.dto.Types;
 import com.tale.exception.TipException;
 import com.tale.init.TaleConst;
-import com.tale.model.Contents;
+import com.tale.model.entity.Contents;
 import com.tale.service.ContentsService;
 import com.tale.service.MetasService;
 import com.tale.utils.TaleUtils;
