+++ src/main/java/com/tale/service/impl/OptionsServiceImpl.java
@@ -5,7 +5,7 @@ import com.blade.ioc.annotation.Inject;
 import com.blade.jdbc.ActiveRecord;
 import com.blade.jdbc.core.Take;
 import com.blade.kit.StringKit;
-import com.tale.model.Options;
+import com.tale.model.entity.Options;
 import com.tale.service.OptionsService;
 
 import java.util.HashMap;
