+++ src/main/java/com/tale/utils/TaleUtils.java
@@ -15,8 +15,8 @@ import com.tale.controller.admin.AttachController;
 import com.tale.extension.Commons;
 import com.tale.extension.Theme;
 import com.tale.init.TaleConst;
-import com.tale.model.Contents;
-import com.tale.model.Users;
+import com.tale.model.entity.Contents;
+import com.tale.model.entity.Users;
 import org.commonmark.Extension;
 import org.commonmark.ext.gfm.tables.TablesExtension;
 import org.commonmark.node.Node;
