+++ src/main/java/com/tale/webhook/BaseWebHook.java
@@ -8,9 +8,9 @@ import com.blade.mvc.hook.Invoker;
 import com.blade.mvc.hook.WebHook;
 import com.blade.mvc.http.Request;
 import com.blade.mvc.http.Response;
-import com.tale.dto.Types;
+import com.tale.model.dto.Types;
 import com.tale.init.TaleConst;
-import com.tale.model.Users;
+import com.tale.model.entity.Users;
 import com.tale.service.UsersService;
 import com.tale.utils.MapCache;
 import com.tale.utils.TaleUtils;
